backend/python/middleware/auth.py
from functools import wraps
from flask import request, jsonify
import jwt
import os

def auth_required(f):
    @wraps(f)
    def decorated(*args, **kwargs):
        token = request.headers.get('Authorization')
        if not token:
            return jsonify({'error': 'No token provided'}), 401
        
        try:
            if token.startswith('Bearer '):
                token = token[7:]
            decoded = jwt.decode(token, os.getenv('JWT_SECRET', 'codecade_secret_key'), algorithms=['HS256'])
            request.user_id = decoded['userId']
            request.user = decoded
        except jwt.ExpiredSignatureError:
            return jsonify({'error': 'Token expired'}), 401
        except jwt.InvalidTokenError:
            return jsonify({'error': 'Invalid token'}), 401
        
        return f(*args, **kwargs)
    return decorated
