backend/python/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SocketIO==5.3.5
python-dotenv==1.0.0
pymongo==4.6.1
bcrypt==4.1.2
PyJWT==2.8.0
gunicorn==21.2.0
